EDIT: Price field, decimal -> Postigive Int

+++ api/model/product.py
@@ -3,7 +3,7 @@ from django.db import models
 class Product(models.Model):
     name = models.CharField(max_length=200)
     description = models.TextField()
-    price = models.DecimalField(max_digits=10, decimal_places=2)
+    price = models.PositiveIntegerField(max_digits=10)
     img_url = models.URLField(max_length=500, null=True, blank=True)
     # stock_quantity = models.PositiveIntegerField()
     created_at = models.DateTimeField(auto_now_add=True)
